Develop

+++ src/main.cpp
@@ -120,10 +120,16 @@ protected:
           Poco::cat(std::string("\n  "), environList.begin(), environList.end()));
     }
 
-    // Prepare for the working directory
+    // Prepare for the working directory and other directories
     Poco::File workDir(_workDir);
     if (!workDir.exists()) {
       workDir.createDirectories();
+    }
+    if (!_outputFile.empty()) {
+      Utils::makeParents(_outputFile);
+    }
+    if (!_statusFile.empty()) {
+      Utils::makeParents(_statusFile);
     }
 
     // Initialize all related objects.
@@ -167,8 +173,8 @@ protected:
         {
           // This nested scope must exist, such that the child process will not install
           // this signal handler.
-          SignalHandler signalHandler([&executor] {
-            Logger::getLogger().info("Termination signal received, kill the user program ...");
+          SignalHandler signalHandler([&executor] (int signalValue) {
+            Logger::getLogger().info("Termination signal %d received, kill the user program ...", signalValue);
             executor.kill();
           });
           executor.wait();
@@ -227,7 +233,7 @@ protected:
     }
 
     // run command after execution
-    if (!_runAfter.empty() && !SignalHandler::interrupted()) {
+    if (!_runAfter.empty()) {
       // prepare for the command
       ArgList runAfterArgs = {shell, "-c", _runAfter};
       EnvironMap environ(_environ);
@@ -251,11 +257,11 @@ protected:
       // run the command
       ProgramExecutor runAfterExecutor(runAfterArgs, environ, std::string(), false, "Run-after command");
       runAfterExecutor.start();
-      SignalHandler signalHandler([&runAfterExecutor] {
-        Logger::getLogger().info("Termination signal received, kill \"run after\" command.");
+      if (!runAfterExecutor.wait(ML_GRIDENGINE_RUN_AFTER_TIMEOUT_SECONDS * 1000)) {
+        Logger::getLogger().error(
+            "Run-after command did not finish in %d seconds.", ML_GRIDENGINE_RUN_AFTER_TIMEOUT_SECONDS);
         runAfterExecutor.kill();
-      });
-      runAfterExecutor.wait();
+      }
     }
 
     // wait for the persist and callback manager to finish
@@ -267,8 +273,9 @@ protected:
 
       // Now wait for the persistAndCallbackManager to exit.
       {
-        SignalHandler signalHandler([&persistAndCallback] {
-          Logger::getLogger().info("Termination signal received, stop the persist and callback manager ...");
+        SignalHandler signalHandler([&persistAndCallback] (int signalValue) {
+          Logger::getLogger().info(
+              "Termination signal %d received, stop the persist and callback manager ...", signalValue);
           persistAndCallback.interrupt();
         });
         persistAndCallback.wait();
@@ -278,8 +285,8 @@ protected:
 
     // If no exit, wait for termination
     if (_noExit && !SignalHandler::interrupted()) {
-      SignalHandler signalHandler([] {
-        Logger::getLogger().info("Termination signal received.");
+      SignalHandler signalHandler([] (int signalValue) {
+        Logger::getLogger().info("Termination signal %d received.", signalValue);
       });
       signalHandler.wait();
     }
